create cookie parsing function
