+++ node/utils/cookie.ts
@@ -1,8 +1,27 @@
 import * as Cookies from "cookies"
+import { parse } from 'cookie'
+import { keys } from 'ramda'
 
 const isUserLoggedIn = (ctx: Context) => {
   const { vtex: { account } } = ctx
   return !!ctx.cookies.get(`VtexIdclientAutCookie_${account}`)
+}
+
+const parseCookie = (cookie: string) => {
+  const parsed = parse(cookie)
+  const cookieName = keys(parsed)[0]
+  const cookieValue = parsed[cookieName]
+
+  const extraOptions = {
+    path: parsed.path,
+    domain: parsed.domain,
+    expires: parsed.expires ? new Date(parsed.expires) : undefined,
+  }
+  return {
+    name: cookieName,
+    value: cookieValue,
+    options: extraOptions,
+  }
 }
 
 /** Checkout cookie methods */
@@ -15,4 +34,4 @@ const getOrderFormIdFromCookie = (cookies: Cookies) => {
   return cookie && cookie.split('=')[1]
 }
 
-export { isUserLoggedIn, CHECKOUT_COOKIE, checkoutCookieFormat, getOrderFormIdFromCookie }
+export { isUserLoggedIn, CHECKOUT_COOKIE, checkoutCookieFormat, getOrderFormIdFromCookie, parseCookie }
